モデルの更新に小数点まで打ち込めるようにしました

+++ delay_predict/app.py
@@ -6,6 +6,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.ensemble import RandomForestRegressor as RFR
 import init_data
 
+
 # 学習済みモデルを読み込み利用します
 def predict(parameters):
     # モデルを読み込み
@@ -14,12 +15,13 @@ def predict(parameters):
     pred = model.predict(params)
     return pred
 
+
 def update(parameters, label):
     # データ取得
-    data = np.load(file ="data.npy")
-    target = np.load(file ="target.npy")
-    x = np.append(data,[parameters],axis=0)
-    y= np.append(target,[label],axis=0)
+    data = np.load(file="data.npy")
+    target = np.load(file="target.npy")
+    x = np.append(data, [parameters], axis=0)
+    y = np.append(target, [label], axis=0)
     # 訓練データとテストデータに分割
     x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=0)
     model = RFR(n_jobs=1, random_state=0, n_estimators=5)
@@ -32,7 +34,9 @@ def update(parameters, label):
     score = model.score(x_test, y_test)
     return score
 
+
 app = Flask(__name__)
+
 
 # Flaskとwtformsを使い、index.html側で表示させるフォームを構築します。
 class DelayForm(Form):
@@ -55,11 +59,13 @@ class DelayForm(Form):
     # html側で表示するsubmitボタンの表示
     submit = SubmitField("予測")
 
+
 class LabelForm(Form):
-    Label = IntegerField("遅延時間(分)",
-                         [validators.InputRequired("この項目は入力必須です"),
-                          validators.NumberRange(min=0, max=10)])
+    Label = FloatField("遅延時間(分)",
+                       [validators.InputRequired("この項目は入力必須です"),
+                        validators.NumberRange(min=0, max=10)])
     submit = SubmitField("更新")
+
 
 @app.route('/', methods=['GET', 'POST'])
 def predicts():
@@ -78,12 +84,13 @@ def predicts():
             pred = predict(x)
             np.save("parameters", x)
 
-            return render_template('result.html', pred=round(pred[0],1))
+            return render_template('result.html', pred=round(pred[0], 1))
     elif request.method == 'GET':
 
         return render_template('index.html', form=form)
 
-@app.route('/update/', methods = ['GET', 'POST'])
+
+@app.route('/update/', methods=['GET', 'POST'])
 def updates():
     form = LabelForm(request.form)
     if request.method == 'POST':
@@ -92,14 +99,15 @@ def updates():
             return render_template('update.html', form=form)
         else:
             Label = int(request.form["Label"])
-            Parameters = np.load(file ="parameters.npy")
+            Parameters = np.load(file="parameters.npy")
             score = update(Parameters, Label)
 
-            return render_template('update_result.html', score=round(score, 3)*100)
+            return render_template('update_result.html', score=round(score, 3) * 100)
     elif request.method == 'GET':
         return render_template('update.html', form=form)
 
-@app.route('/init/', methods = ['GET', 'POST'])
+
+@app.route('/init/', methods=['GET', 'POST'])
 def inits():
     if request.method == 'POST':
         init_data.init_data()
@@ -107,12 +115,14 @@ def inits():
     elif request.method == 'GET':
         return render_template('init.html')
 
+
 @app.route('/list/')
 def list():
-    data = np.load(file ="data.npy")
-    target = np.load(file ="target.npy")
-    x = zip(data,target)
+    data = np.load(file="data.npy")
+    target = np.load(file="target.npy")
+    x = zip(data, target)
     return render_template('list.html', data=x)
+
 
 if __name__ == "__main__":
     app.run()
